feat: add useCustomActionSheet prop to provider
